mcp/requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Async I/O
aiofiles>=23.0.0
aiosqlite>=0.19.0

# File integrity and hashing
xxhash>=3.4.0
blake3>=0.4.0

# File system monitoring
watchdog>=3.0.0

# AI/ML features (optional but recommended)
sentence-transformers>=2.2.0
torch>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Text processing
tiktoken>=0.5.0
pygments>=2.16.0

# Utilities
PyYAML>=6.0.0